refactor: move milk tea popup logic
